refactor(proj): 将杨桃的创建所在行标记从int改为byte

+++ pvzdll/ProjectileEvents.cpp
@@ -5,7 +5,7 @@ bool onPlantAddProjectile(MyPlant plant, MyProjectile proj, MyZombie zombie)
 {
 	proj.ParentID = plant.Id;
 	if (plant.Type == SeedType::Starfruit)
-		proj.OriginalRow = plant.Row;
+		proj.OriginalRow = (byte)plant.Row;
 	return true;
 }
